requirements.txt
fastapi
uvicorn
jinja2
pymongo
python-multipart
python-dotenv
